+++ afs/afs-storage-api/src/test/java/com/powsybl/afs/storage/NodeInfoTest.java
@@ -36,7 +36,7 @@ public class NodeInfoTest {
                                          .setDouble("d1", 1d)
                                          .setInt("i1", 2)
                                          .setBoolean("b1", true),
-                new NodeAccessRights().setGroupRights("group1", (short) 2).setGroupRights("group2", (short) 4).setUserRights("user1", (short) 6));
+                new NodeAccessRights().setGroupRights("group1", 2).setGroupRights("group2", 4).setUserRights("user1", 6));
         NodeInfo info2 = objectMapper.readValue(objectMapper.writeValueAsString(info), NodeInfo.class);
         assertEquals(info, info2);
         info.setVersion(1);
